Refactored, added package routes and modules task_route.py and goal_route.py
