+++ app/__init__.py
@@ -17,11 +17,12 @@ def create_app(test_config=None):
     if test_config is None:
         app.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get(
             "SQLALCHEMY_DATABASE_URI")
+        # app.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get(
+        #     "RENDER_DATABASE_URL")
     else:
         app.config["TESTING"] = True
         app.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get(
             "SQLALCHEMY_TEST_DATABASE_URI")
-
     # Import models here for Alembic setup
     from app.models.task import Task
     from app.models.goal import Goal
@@ -30,9 +31,9 @@ def create_app(test_config=None):
     migrate.init_app(app, db)
 
     # Register Blueprints here
-    from .routes import task_bp
+    from app.routes.task_route import task_bp
     app.register_blueprint(task_bp)
-    from .routes import goal_bp
+    from app.routes.goal_route import goal_bp
     app.register_blueprint(goal_bp)
 
     return app
