+++ tests/test_wave_03.py
@@ -45,7 +45,10 @@ def test_mark_complete_on_incomplete_task(client, one_task):
 # @pytest.mark.skip(reason="No way to test this feature yet")
 def test_mark_incomplete_on_complete_task(client, completed_task):
     # Act
-    response = client.patch("/tasks/1/mark_incomplete")
+    with patch("requests.post") as mock_get:
+        mock_get.return_value.status_code = 200
+        response = client.patch("/tasks/1/mark_incomplete")
+
     response_body = response.get_json()
 
     # Assert
@@ -102,7 +105,9 @@ def test_mark_complete_on_completed_task(client, completed_task):
 # @pytest.mark.skip(reason="No way to test this feature yet")
 def test_mark_incomplete_on_incomplete_task(client, one_task):
     # Act
-    response = client.patch("/tasks/1/mark_incomplete")
+    with patch("requests.post") as mock_get:
+        mock_get.return_value.status_code = 200
+        response = client.patch("/tasks/1/mark_incomplete")
     response_body = response.get_json()
 
     # Assert
